refact: isolando componentes relacionados aos states dos files
