+++ src/app.tsx
@@ -1,116 +1,18 @@
-import { useRef, useState, useEffect, ChangeEvent, MouseEvent } from 'react'
 import styled from 'styled-components/macro'
-import { v4 as uuidv4 } from 'uuid'
-import { File } from 'resources/files/types'
 import { Sidebar } from 'pages/sidebar'
 import { Content } from 'pages/content'
+import { useFiles } from 'pages/files/use-files'
 
 export function App () {
-  const inputRef = useRef<HTMLInputElement>(null)
-  const [files, setFiles] = useState<File[]>([])
-
-  useEffect(() => {
-    let timer: ReturnType<typeof setTimeout>
-
-    function updateStatus () {
-      const file = files.find(file => file.active === true)
-
-      if (!file || file.status !== 'editing') {
-        return
-      }
-
-      timer = setTimeout(() => {
-        setFiles(files => files.map(file => {
-          if (file.active) {
-            return {
-              ...file,
-              status: 'saving',
-            }
-          }
-
-          return file
-        }))
-
-        setTimeout(() => {
-          setFiles(files => files.map(file => {
-            if (file.active) {
-              return {
-                ...file,
-                status: 'saved',
-              }
-            }
-
-            return file
-          }))
-        }, 300)
-      }, 300)
-    }
-
-    updateStatus()
-
-    return () => clearTimeout(timer)
-  }, [files])
-
-  const handleCreateNewFile = () => {
-    inputRef.current?.focus()
-
-    setFiles(files => files
-      .map(file => ({
-        ...file,
-        active: false,
-      }))
-      .concat({
-        id: uuidv4(),
-        name: 'Sem título',
-        content: '',
-        active: true,
-        status: 'saved',
-      }),
-    )
-  }
-
-  const handleUpdateFileName = (id: string) => (e: ChangeEvent<HTMLInputElement>) => {
-    setFiles(files => files.map(file => {
-      if (file.id === id) {
-        return {
-          ...file,
-          name: e.target.value,
-          status: 'editing',
-        }
-      }
-
-      return file
-    }))
-  }
-
-  const handleUpdateFileContent = (id: string) => (e: ChangeEvent<HTMLTextAreaElement>) => {
-    setFiles(files => files.map(file => {
-      if (file.id === id) {
-        return {
-          ...file,
-          content: e.target.value,
-          status: 'editing',
-        }
-      }
-
-      return file
-    }))
-  }
-
-  const handleSelectFile = (id: string) => (e: MouseEvent) => {
-    e.preventDefault()
-
-    inputRef.current?.focus()
-
-    setFiles(files => files.map(file => ({
-      ...file,
-      active: file.id === id,
-    })))
-  }
-
-  const handleRemoveFile = (id: string) => {
-    setFiles(files => files.filter(file => file.id !== id))
-  }
+  const {
+    files,
+    inputRef,
+    handleCreateNewFile,
+    handleSelectFile,
+    handleRemoveFile,
+    handleUpdateFileContent,
+    handleUpdateFileName,
+  } = useFiles()
 
   return (
     <Main>
